+++ backends/slack/src/main/kotlin/com/gatehill/slackgateway/backend/slack/model/SlackChannel.kt
@@ -1,0 +1,7 @@
+package com.gatehill.slackgateway.backend.slack.model
+
+interface SlackChannel {
+    val id: String
+    val name: String
+    val members: List<String>
+}