+++ backends/slack/src/main/kotlin/com/gatehill/slackgateway/backend/slack/model/SlackChannelsCreateResponse.kt
@@ -1,0 +1,5 @@
+package com.gatehill.slackgateway.backend.slack.model
+
+interface SlackChannelsCreateResponse {
+    val channel: SlackChannel
+}