+++ backends/slack/src/main/kotlin/com/gatehill/slackgateway/backend/slack/model/SlackPublicChannel.kt
@@ -2,11 +2,8 @@ package com.gatehill.slackgateway.backend.slack.model
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 
-/**
- * This is actually a private channel.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
-data class SlackGroup(
+data class SlackPublicChannel(
     override val id: String,
     override val name: String,
     override val members: List<String>