+++ backends/slack/src/main/kotlin/com/gatehill/slackgateway/backend/slack/service/SlackOperationsService.kt
@@ -2,14 +2,20 @@ package com.gatehill.slackgateway.backend.slack.service
 
 import com.gatehill.slackgateway.backend.slack.config.SlackSettings
 import com.gatehill.slackgateway.backend.slack.exception.SlackErrorResponseException
+import com.gatehill.slackgateway.backend.slack.model.ChannelsCreateResponse
+import com.gatehill.slackgateway.backend.slack.model.ChannelsListResponse
 import com.gatehill.slackgateway.backend.slack.model.GroupsCreateResponse
 import com.gatehill.slackgateway.backend.slack.model.GroupsListResponse
+import com.gatehill.slackgateway.backend.slack.model.SlackChannel
+import com.gatehill.slackgateway.backend.slack.model.SlackChannelsCreateResponse
 import com.gatehill.slackgateway.backend.slack.model.SlackGroup
+import com.gatehill.slackgateway.backend.slack.model.SlackPublicChannel
 import com.gatehill.slackgateway.backend.slack.model.SlackUser
 import com.gatehill.slackgateway.backend.slack.model.SlackUserGroup
 import com.gatehill.slackgateway.backend.slack.model.UserGroupsListResponse
 import com.gatehill.slackgateway.backend.slack.model.UserGroupsUsersListResponse
 import com.gatehill.slackgateway.backend.slack.model.UsersListResponse
+import com.gatehill.slackgateway.model.ChannelType
 import com.google.common.cache.CacheBuilder
 import com.google.common.cache.CacheLoader
 import org.apache.logging.log4j.LogManager
@@ -53,59 +59,84 @@ class SlackOperationsService @Inject constructor(
     private fun fetchUsers(): List<SlackUser> {
         logger.debug("Fetching all users")
 
-        val reply = slackApiService.invokePaginatedSlackCommand<SlackUser, UsersListResponse>(commandName = "users.list")
+        val reply =
+            slackApiService.invokePaginatedSlackCommand<SlackUser, UsersListResponse>(commandName = "users.list")
         return reply.members
     }
 
     private fun fetchUserGroups(): List<SlackUserGroup> {
         logger.debug("Fetching all user groups")
-
         val reply = slackApiService.invokeSlackCommand<UserGroupsListResponse>(
             commandName = "usergroups.list",
             method = SlackApiService.HttpMethod.GET
         )
-
         return reply.usergroups
     }
 
-    internal fun listPrivateChannels(): List<SlackGroup> {
-        logger.debug("Listing private channels")
+    internal fun listChannels(channelType: ChannelType): List<SlackChannel> = when (channelType) {
+        ChannelType.PRIVATE -> listPrivateChannels()
+        ChannelType.PUBLIC -> listPublicChannels()
+    }
 
-        val reply = slackApiService.invokePaginatedSlackCommand<SlackGroup, GroupsListResponse>(commandName = "groups.list")
+    private fun listPrivateChannels(): List<SlackGroup> {
+        logger.debug("Listing private channels")
+        val reply = slackApiService.invokePaginatedSlackCommand<SlackGroup, GroupsListResponse>(
+            commandName = "groups.list"
+        )
         return reply.groups ?: emptyList()
     }
 
-    internal fun createPrivateChannel(channelName: String): SlackGroup {
-        logger.debug("Creating channel: $channelName")
+    private fun listPublicChannels(): List<SlackPublicChannel> {
+        logger.debug("Listing public channels")
+        val reply = slackApiService.invokePaginatedSlackCommand<SlackPublicChannel, ChannelsListResponse>(
+            commandName = "channels.list"
+        )
+        return reply.channels ?: emptyList()
+    }
 
-        val reply: GroupsCreateResponse = try {
-            slackApiService.invokeSlackCommand(
-                commandName = "groups.create",
-                params = mapOf(
-                    "name" to channelName,
-                    "validate" to "true"
+    internal fun createChannel(channelName: String, channelType: ChannelType): SlackChannel {
+        logger.debug("Creating $channelType channel: $channelName")
+
+        val reply: SlackChannelsCreateResponse = try {
+            when (channelType) {
+                ChannelType.PRIVATE -> slackApiService.invokeSlackCommand<GroupsCreateResponse>(
+                    commandName = "groups.create",
+                    params = mapOf(
+                        "name" to channelName,
+                        "validate" to "true"
+                    )
                 )
-            )
+                ChannelType.PUBLIC -> slackApiService.invokeSlackCommand<ChannelsCreateResponse>(
+                    commandName = "channels.create",
+                    params = mapOf(
+                        "name" to channelName,
+                        "validate" to "true"
+                    )
+                )
+            }
+
         } catch (e: SlackErrorResponseException) {
             // specifically handle case where group exists but bot cannot see it
             if (e.errorResponse?.error == "name_taken") {
                 throw IllegalStateException(
-                    "Unable to create a private channel named '$channelName' as one already exists. This is usually because the Slack token you are using does not have permission to access an existing channel. Add permission to access the '$channelName' channel to this token and try again."
+                    "Unable to create a $channelType channel named '$channelName' as one already exists. This is usually because the Slack token you are using does not have permission to access an existing channel. Add permission to access the '$channelName' channel to this token and try again."
                 )
             } else {
                 throw e
             }
         }
 
-        logger.debug("Create channel response: $reply")
-
-        logger.debug("Channel $channelName created")
-        return reply.group
+        logger.debug("Create $channelType channel response: $reply")
+        logger.info("Channel $channelName ($channelType) created")
+        return reply.channel
     }
 
-    internal fun inviteToPrivateChannel(channel: SlackGroup, memberId: String) {
+    internal fun inviteToChannel(channel: SlackChannel, channelType: ChannelType, memberId: String) {
         slackApiService.invokeSlackCommand<Map<String, Any>>(
-            commandName = "groups.invite",
+            commandName = when (channelType) {
+                ChannelType.PRIVATE -> "groups.invite"
+                ChannelType.PUBLIC -> "channels.invite"
+            },
             params = mapOf(
                 "channel" to channel.id,
                 "user" to memberId