+++ core/src/main/kotlin/com/gatehill/slackgateway/service/OutboundMessageService.kt
@@ -1,11 +1,13 @@
 package com.gatehill.slackgateway.service
 
+import com.gatehill.slackgateway.model.ChannelType
+
 /**
  * Handles outbound messages.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 interface OutboundMessageService {
-    fun forward(raw: String)
-    fun forward(message: Map<String, *>)
+    fun forward(raw: String, channelType: ChannelType)
+    fun forward(message: Map<String, *>, channelType: ChannelType)
 }