+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Adds support for creating and posting to public channels.
+
 ## [0.2.0] - 2019-08-08
 ### Added
 - Adds paging support for listing users and private channels.